CCS/requirements.txt
pyjwt
graphql-core-next
toolz
pymongo
aiodocker
aioredis
python-multipart
fastapi
influxdb
kubernetes
numpy
keras==2.3.1
sklearn
grpcio
cramjam
protobuf
tensorflow==2.0.0
PyWavelets
seglearn
pandas
EMD-signal